+++ genhooks/doc.go
@@ -1,0 +1,2 @@
+// Package genhooks provides a set of entgo hooks for generating files from the ent schema.
+package genhooks
