+++ genhooks/gen.go
@@ -1,4 +1,4 @@
-package entx
+package genhooks
 
 import (
 	"embed"
