+++ genhooks/templates/graph.tpl
