+++ genhooks/templates/query.tpl
