+++ vanilla/_example/schema/ent.graphql
@@ -308,6 +308,10 @@ type Query {
   ): [Node]!
 }
 """
+The builtin Time type
+"""
+scalar Time
+"""
 UpdateOrgMembershipInput is used for update OrgMembership object.
 Input was generated by ent.
 """
